refactor(gateway): gate all distributed writes at the repository choke point
The concurrency bound lived in the sync service and wrapped only createOrUpdate,
so the undeploy path (updateAll) — which writes directly through the repository —
was left unbounded: a mass undeploy could still overflow the Redis waiting queue.

Move the bound into a dedicated DistributedWriteGate in the repository, wrapping
createOrUpdateKey, through which every distributed write (createOrUpdate and
updateAll) funnels. This bounds all write paths uniformly, co-locates the gate
with the Redis max-waiting-handlers it protects, and leaves the sync service with
just its per-deployable fan-out cap.

APIM-14672
