+++ gravitee-apim-gateway/gravitee-apim-gateway-services/gravitee-apim-gateway-services-sync/src/test/java/io/gravitee/gateway/services/sync/process/distributed/service/DefaultDistributedSyncServiceTest.java
@@ -65,10 +65,6 @@ import io.reactivex.rxjava3.core.Completable;
 import io.reactivex.rxjava3.core.Maybe;
 import java.util.List;
 import java.util.Set;
-import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
-import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.stream.IntStream;
 import org.junit.jupiter.api.BeforeEach;
@@ -415,150 +411,6 @@ class DefaultDistributedSyncServiceTest {
             // The failure flag is reset: the next cycle can store its state again
             when(distributedSyncStateRepository.createOrUpdate(any())).thenReturn(Completable.complete());
             cut.storeState(1L, 2L).test().assertComplete();
-        }
-
-        @Test
-        void should_bound_concurrent_writes_across_parallel_distributions() {
-            // WRITE_MAX_CONCURRENCY only caps a single deployable; the deployer distributes many in parallel.
-            // The shared gate must bound the AGGREGATE in-flight writes, not just each call individually.
-            cut.writeGate = new DefaultDistributedSyncService.WriteGate(8);
-
-            AtomicInteger current = new AtomicInteger();
-            AtomicInteger peak = new AtomicInteger();
-            // Count via doOnSubscribe/doOnTerminate (which run before the terminal is propagated) so the
-            // gauge reflects true concurrency; doFinally would over-count as the freed permit is handed to
-            // the next write before the finishing write's decrement runs.
-            when(distributedEventRepository.createOrUpdate(any())).thenAnswer(invocation ->
-                Completable.timer(20, TimeUnit.MILLISECONDS)
-                    .doOnSubscribe(disposable -> peak.accumulateAndGet(current.incrementAndGet(), Math::max))
-                    .doOnTerminate(current::decrementAndGet)
-            );
-
-            List<Completable> distributions = IntStream.range(0, 100)
-                .mapToObj(i ->
-                    cut.distributeIfNeeded(SingleSubscriptionDeployable.builder().subscription(subscription("subscription-" + i)).build())
-                )
-                .toList();
-
-            Completable.merge(distributions).test().awaitDone(10, TimeUnit.SECONDS).assertComplete();
-
-            verify(distributedEventRepository, times(100)).createOrUpdate(any());
-            // Exactly 8: the gate must both cap concurrency at 8 AND actually reach it (100 writes are
-            // submitted at once), so an over-serializing regression such as a permit leak is caught here.
-            assertThat(peak.get()).isEqualTo(8);
-        }
-
-        @Test
-        void write_gate_should_recover_the_permit_when_a_holder_is_disposed() {
-            var gate = new DefaultDistributedSyncService.WriteGate(1);
-
-            var holder = gate.runGated(Completable.never()).test();
-            var parked = gate.runGated(Completable.complete()).test();
-            parked.assertNotComplete(); // the single permit is held, so this run parks
-
-            holder.dispose(); // frees the permit -> handed to the parked run
-            parked.assertComplete();
-
-            // permit fully returned to the pool afterwards: a fresh run completes immediately
-            gate.runGated(Completable.complete()).test().assertComplete();
-        }
-
-        @Test
-        void write_gate_should_not_release_a_permit_when_a_parked_waiter_is_disposed() {
-            var gate = new DefaultDistributedSyncService.WriteGate(1);
-
-            var holder = gate.runGated(Completable.never()).test(); // holds the only permit
-            var parked = gate.runGated(Completable.never()).test(); // parks, never granted
-
-            parked.dispose(); // disposing the WAITER must not free the holder's permit
-
-            var next = gate.runGated(Completable.complete()).test();
-            next.assertNotComplete(); // gate still full, holder keeps its permit
-
-            holder.dispose(); // now a permit frees and is handed to next
-            next.assertComplete();
-        }
-
-        @Test
-        void write_gate_should_return_every_permit_under_concurrent_completions_errors_and_disposals() throws InterruptedException {
-            var gate = new DefaultDistributedSyncService.WriteGate(4);
-            int threads = 8;
-            int opsPerThread = 200;
-            ExecutorService pool = Executors.newFixedThreadPool(threads);
-            CountDownLatch done = new CountDownLatch(threads);
-            for (int t = 0; t < threads; t++) {
-                final int seed = t;
-                pool.submit(() -> {
-                    try {
-                        for (int i = 0; i < opsPerThread; i++) {
-                            int kind = (seed + i) % 3;
-                            Completable work = switch (kind) {
-                                case 0 -> Completable.complete();
-                                case 1 -> Completable.error(new RuntimeException("boom"));
-                                default -> Completable.timer(2, TimeUnit.MILLISECONDS);
-                            };
-                            var observer = gate.runGated(work).test();
-                            if (kind == 2 && i % 2 == 0) {
-                                observer.dispose(); // race a disposal against the in-flight write
-                            } else {
-                                observer.awaitDone(5, TimeUnit.SECONDS);
-                            }
-                        }
-                    } finally {
-                        done.countDown();
-                    }
-                });
-            }
-            assertThat(done.await(60, TimeUnit.SECONDS)).isTrue();
-            pool.shutdownNow();
-
-            // Every permit must have been returned, whatever the interleaving of completes/errors/disposals.
-            int inFlight = gate.inFlight();
-            for (int i = 0; i < 200 && inFlight != 0; i++) {
-                Thread.sleep(10);
-                inFlight = gate.inFlight();
-            }
-            assertThat(inFlight).isZero();
-            // ...and the gate still grants afterwards.
-            gate.runGated(Completable.complete()).test().awaitDone(5, TimeUnit.SECONDS).assertComplete();
-        }
-
-        @Test
-        void should_not_drop_events_when_redis_waiting_queue_would_overflow() {
-            // Reproduces APIM-14672: the Redis client rejects writes with "Redis waiting queue is full" once
-            // too many are in-flight, silently dropping subscription/api-key events during a bulk sync. The
-            // global bound must keep in-flight under that ceiling so every event is still written.
-            int waitingQueueCeiling = 8;
-            cut.writeGate = new DefaultDistributedSyncService.WriteGate(waitingQueueCeiling);
-
-            AtomicInteger current = new AtomicInteger();
-            when(distributedEventRepository.createOrUpdate(any())).thenAnswer(invocation ->
-                Completable.defer(() -> {
-                    if (current.incrementAndGet() > waitingQueueCeiling) {
-                        current.decrementAndGet();
-                        return Completable.error(new RuntimeException("Redis waiting queue is full"));
-                    }
-                    // Decrement before the terminal is propagated (doOnTerminate, not doFinally) so the count
-                    // is accurate when the freed permit is immediately handed to the next queued write.
-                    return Completable.timer(20, TimeUnit.MILLISECONDS).doOnTerminate(current::decrementAndGet);
-                })
-            );
-
-            List<Completable> distributions = IntStream.range(0, 200)
-                .mapToObj(i ->
-                    cut.distributeIfNeeded(SingleSubscriptionDeployable.builder().subscription(subscription("subscription-" + i)).build())
-                )
-                .toList();
-
-            Completable.merge(distributions).test().awaitDone(10, TimeUnit.SECONDS).assertComplete();
-
-            verify(distributedEventRepository, times(200)).createOrUpdate(any());
-        }
-
-        private Subscription subscription(final String id) {
-            Subscription subscription = new Subscription();
-            subscription.setId(id);
-            return subscription;
         }
     }
 }
